requirements.txt
pandas==2.1.4
Pillow==10.1.0
Pillow==10.1.0
Requests==2.31.0
scikit_surprise==1.1.3
st_click_detector==0.1.3
